requirements.txt
fastapi
uvicorn
langchain-google-genai
python-dotenv
langchain-core
langchain-community